node_metrics::api::node_validator::v0

Module cdn

Structs

BroadcastRollCallTask

BroadcastRollCallTask represents a task that is responsible for broadcasting a RollCallRequest to the CDN network.

CdnReceiveMessagesTask

CdnReceiveMessagesTask represents a task that is responsible for receiving messages from the CDN network and processing them. This task is primarily concerned with recording responses to RollCall requests, and forwarding any discovered public API URLs to the URL sender.

Traits

ConnectedNetworkConsumer

ConnectedNetworkConsumer represents a trait that splits up a portion of the ConnectedNetwork trait, so that the consumer only needs to be aware of the wait_for_ready and recv_message functions.

ConnectedNetworkPublisher

ConnectedNetworkPublisher represents a trait that splits up a portion of the ConnectedNetwork trait, so that the consumer only needs to be aware of the wait_for_ready and broadcast_message functions.