node_metrics::api::node_validator::v0::cdn
Trait ConnectedNetworkConsumer

pub trait ConnectedNetworkConsumer<K> {
    // Required methods
    fn wait_for_ready<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn recv_message<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>, NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

ConnectedNetworkConsumer represents a trait that splits up a portion of the ConnectedNetwork trait, so that the consumer only needs to be aware of the wait_for_ready and recv_message functions.

Required Methods

fn wait_for_ready<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

[wait_for_ready] will not return until the network is ready to be utilized.

fn recv_message<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>, NetworkError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

[recv_message] will return a list of messages that have been received from the network.

Errors

All errors are expected to be network related.