node_metrics::api::node_validator::v0::cdn
Trait ConnectedNetworkPublisher

pub trait ConnectedNetworkPublisher<K> {
    // Required methods
    fn wait_for_ready<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn broadcast_message<'life0, 'async_trait>(
        &'life0 self,
        message: Vec<u8>,
        topic: Topic,
        broadcast_delay: BroadcastDelay,
    ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

ConnectedNetworkPublisher represents a trait that splits up a portion of the ConnectedNetwork trait, so that the consumer only needs to be aware of the wait_for_ready and broadcast_message functions.

Required Methods

fn wait_for_ready<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

[wait_for_ready] will not return until the network is ready to be utilized.

fn broadcast_message<'life0, 'async_trait>( &'life0 self, message: Vec<u8>, topic: Topic, broadcast_delay: BroadcastDelay, ) -> Pin<Box<dyn Future<Output = Result<(), NetworkError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

[broadcast_message] will broadcast the given message to some subset of nodes in the network based on the given topic.

This is a blocking operation.