node_metrics::api::node_validator::v0

Module create_node_validator_api

Structs

HotShotEventProcessingTask

HotShotEventProcessingTask is a task that is capable of processing events that are coming in from a HotShot event stream. This task will keep an eye out for ExternalMessageReceived events that can be decoded as a RollCallResponse. When a RollCallResponse is received, the public API URL of the node that sent the message will be sent to the provided url_sender.

NodeValidatorAPI

NodeValidatorConfig

ProcessExternalMessageHandlingTask

ProcessExternalMessageHandlingTask is a task that is capable of processing external messages that are coming in from an external message stream. This task will keep an eye out for ExternalMessageReceived events that can be decoded as a RollCallResponse. When a RollCallResponse is received, the public API URL of the node that sent the message will be sent to the provided url_sender.

RollCallInfo

Information about a node that is used in a roll call response

Enums

CreateNodeValidatorProcessingError

ExternalMessage

An external message that can be sent to or received from a node

Functions

create_node_validator_processing

create_node_validator_processing is a function that creates a node validator processing environment. This function will create a number of tasks that will be responsible for processing the data streams that are coming in from the various sources. This function will also create the data state that will be used to store the state of the network.