node_metrics::api::node_validator::v0

Function node_identity_from_scrape

pub fn node_identity_from_scrape(scrape: Scrape) -> Option<NodeIdentity>

node_identity_from_scrape creates a NodeIdentity from a [Scrape]. It expects the [Scrape] to contain the necessary information to populate the NodeIdentity. If the [Scrape] doesn’t contain the necessary information to populate the NodeIdentity, then it will return None.