node_metrics::api::node_validator::v0

Trait LeafAndBlockPairStream

pub trait LeafAndBlockPairStream: Stream<Item = LeafAndBlock<SeqTypes>> { }

LeafAndBlockPairStream is a trait that represents a stream of [Leaf1QueryData]s and [BlockQueryData]s. This acts as a simple type declaration for quick reference.