Trait node_metrics::api::node_validator::v0::LeafStreamRetriever

pub trait LeafStreamRetriever: Send {
    type Item;
    type ItemError: Error + Send;
    type Error: Error + Send;
    type Stream: Stream<Item = Result<Self::Item, Self::ItemError>> + Send + Unpin;
    type Future: Future<Output = Result<Self::Stream, Self::Error>> + Send;

    // Required method
    fn retrieve_stream(&self, current_block_height: Option<u64>) -> Self::Future;
}

LeafStreamRetriever is a general trait that allows for the retrieval of a list of Leaves from a source. The specific implementation doesn’t care about the source, only that it is able to retrieve a stream of Leaves.

This allows us to swap the implementation of the LeafStreamRetriever for testing purposes, or for newer sources in the future.

Required Associated Types

type Item

type ItemError: Error + Send

type Error: Error + Send

type Stream: Stream<Item = Result<Self::Item, Self::ItemError>> + Send + Unpin

type Future: Future<Output = Result<Self::Stream, Self::Error>> + Send

Required Methods

fn retrieve_stream(&self, current_block_height: Option<u64>) -> Self::Future

[retrieve_stream] retrieves a stream of [Leaf]s from the source. It expects the current block height to be provided so that it can determine the starting block height to retrieve the stream of [Leaf]s from.

It should check the current height of the chain so that it only needs to retrieve the number of older blocks that are needed, instead of starting from the beginning of time.