node_metrics::api::node_validator::v0

Trait SurfDiscoAvailabilityAPIPathResolver

pub trait SurfDiscoAvailabilityAPIPathResolver {
    // Required method
    fn resolve_path_for_height(&self, height: u64) -> String;
}

SurfDiscoAvailabilityAPIPathResolver is a trait that allows for the specification of a sub path to the base URL that will resolve in a URL to point to the correct endpoint for the desired Stream type.

Many streams in the Availability API have a path that is based on the specific type of data you are wanting to stream, and the block height to start retrieving that data for. This trait allows for us to abstract out these endpoints.

Required Methods

fn resolve_path_for_height(&self, height: u64) -> String

[resolve_path_for_height] resolves the path for the given height. It is expected that the path will be appended to the base URL to create a full URL that can be used to connect to the stream.