node_metrics::service
Module client_state

Structs

ClientState

ClientState represents the service state of the connected clients. It maintains and represents the connected clients, and their subscriptions.

ClientThreadState

ClientThreadState represents the state of all of the active client connections connected to the service. This state governs which clients are connected, and what subscriptions they have setup.

InternalClientMessageProcessingTask

InternalClientMessageProcessingTask represents an async task for InternalClientMessages, and making the appropriate updates to the ClientThreadState and DataState.

ProcessDistributeBlockDetailHandlingTask

ProcessDistributeBlockDetailHandlingTask represents an async task for processing the incoming [BlockDetail] and distributing them to all subscribed clients.

ProcessDistributeNodeIdentityHandlingTask

ProcessDistributeNodeIdentityHandlingTask represents an async task for processing the incoming NodeIdentity and distributing them to all subscribed clients.

ProcessDistributeVotersHandlingTask

ProcessDistributeVotersHandlingTask represents an async task for processing the incoming [BitVec] and distributing them to all subscribed clients.

Enums

HandleConnectedError

HandleConnectedError represents the scope of errors that can be returned from the handle_client_message_connected function.

HandleRequestBlocksSnapshotsError

HandleRequestBlocksSnapshotsError represents the scope of errors that can be returned from the handle_client_message_request_blocks_snapshot function.

HandleRequestHistogramSnapshotError

HandleRequestHistogramSnapshotError represents the scope of errors that can be returned from the handle_client_message_request_histogram_snapshot function.

HandleRequestNodeIdentitySnapshotError

HandleRequestNodeIdentitySnapshotError represents the scope of errors that can be returned from the handle_client_message_request_node_identity_snapshot function.

HandleRequestVotersSnapshotError

ProcessClientMessageError

ProcessClientMessageError represents the scope of errors that can be returned from the process_client_message function.

Functions

clone_block_detail

clone_block_detail is a utility function that clones a [BlockDetail] instance.

drop_client_client_thread_state_write_guard 🔒

drop_client_client_thread_state_write_guard is a utility function for cleaning up the ClientThreadState

drop_client_no_lock_guard 🔒

drop_client_no_lock_guard is a utility function for cleaning up the ClientThreadState when a client is detected as disconnected.

drop_failed_client_sends 🔒

drop_failed_client_sends is a function that will drop all of the failed client sends from the client thread state.

handle_client_message_connected

handle_client_message_connected is a function that processes the client message to connect a client to the service.

handle_client_message_disconnected

handle_client_message_disconnected is a function that processes the client message to disconnect a client from the service.

handle_client_message_request_blocks_snapshot

handle_client_message_request_blocks_snapshot is a function that processes the client message request for a blocks snapshot.

handle_client_message_request_histogram_snapshot

handle_client_message_request_histogram_snapshot is a function that processes the client message request for a histogram snapshot.

handle_client_message_request_node_identity_snapshot

handle_client_message_request_node_identity_snapshot is a function that processes the client message request for a node identity snapshot.

handle_client_message_request_voters_snapshot

handle_client_message_request_voters_snapshot is a function that processes the client message request for a voters snapshot.

handle_client_message_subscribe_latest_block

handle_client_message_subscribe_latest_block is a function that processes the client message to subscribe to the latest block stream.

handle_client_message_subscribe_node_identity

handle_client_message_subscribe_node_identity is a function that processes the client message to subscribe to the node identity stream.

handle_client_message_subscribe_voters

handle_client_message_subscribe_voters is a function that processes the client message to subscribe to the voters bitvecs.

handle_received_block_detail 🔒

handle_received_block_detail is a function that processes received Block details and will attempt to distribute the message to all of the clients that are subscribed to the latest block stream.

handle_received_node_identity 🔒

handle_received_node_identity is a function that processes received NodeIdentity and will attempt to distribute the message to all of the clients that are subscribed to the node identity stream.

handle_received_voters 🔒

handle_received_voters is a function that processes received voters and will attempt to distribute the message to all of the clients that are subscribed to the voters stream.

process_client_message

process_client_message is a function that processes the client message and processes the message accordingly.