node_metrics::service::client_state

Function handle_received_block_detail

async fn handle_received_block_detail<K>(
    client_thread_state: Arc<RwLock<ClientThreadState<K>>>,
    block_detail: BlockDetail<SeqTypes>,
)
where
    K: Sink<ServerMessage, Error = SendError> + Clone + Unpin,

handle_received_block_detail is a function that processes received Block details and will attempt to distribute the message to all of the clients that are subscribed to the latest block stream.