node_metrics::service::client_state
Function handle_received_node_identity

async fn handle_received_node_identity<K>(
    client_thread_state: Arc<RwLock<ClientThreadState<K>>>,
    node_identity: NodeIdentity,
)
where
    K: Sink<ServerMessage, Error = SendError> + Clone + Unpin,

handle_received_node_identity is a function that processes received NodeIdentity and will attempt to distribute the message to all of the clients that are subscribed to the node identity stream.