node_metrics::service::client_state

Function process_client_message

pub async fn process_client_message<K>(
    message: InternalClientMessage<K>,
    data_state: Arc<RwLock<DataState>>,
    client_thread_state: Arc<RwLock<ClientThreadState<K>>>,
) -> Result<(), ProcessClientMessageError>
where
    K: Sink<ServerMessage, Error = SendError> + Clone + Unpin,

process_client_message is a function that processes the client message and processes the message accordingly.

The DataState is provided and is used only as a Read lock to distribute the current state of the system to the clients upon request.

The ClientThreadState is provided as it needs to be updated with new subscriptions / new connections depending on the incoming InternalClientMessage