node_metrics::service

Module data_state

Re-exports

pub use location_details::LocationDetails;

pub use node_identity::NodeIdentity;

Modules

location_details

node_identity

Structs

DataState

DataState represents the state of the data that is being stored within the service.

ProcessLeafStreamTask

ProcessLeafStreamTask represents the task that is responsible for processing a stream of incoming [Leaf]s.

ProcessNodeIdentityStreamTask

ProcessNodeIdentityStreamTask represents the task that is responsible for processing a stream of incoming NodeIdentitys and updating the DataState with the new information.

Enums

ProcessLeafError

ProcessLeafError represents the error that can occur when processing a [Leaf].

ProcessNodeIdentityError

ProcessNodeIdentityError represents the error that can occur when processing a NodeIdentity.

Constants

MAX_HISTORY 🔒

MAX_HISTORY represents the last N records that are stored within the DataState structure for the various different sample types.

Functions

create_block_detail_from_leaf

create_block_detail_from_leaf is a helper function that will build a [BlockDetail] from the reference to [Leaf].

process_incoming_leaf 🔒

process_incoming_leaf is a helper function that will process an incoming [Leaf] and update the DataState with the new information. Additionally, the block that is contained within the [Leaf] will be computed into a [BlockDetail] and sent to the [Sink] so that it can be processed for real-time considerations.

process_incoming_node_identity 🔒

process_incoming_node_identity is a helper function that will process an incoming NodeIdentity and update the DataState with the new information. Additionally, the NodeIdentity will be sent to the [Sink] so that it can be processed for real-time considerations.