node_metrics::api::node_validator
Module v0

Modules

cdn

create_node_validator_api

Structs

HotshotQueryServiceLeafStreamRetriever

HotshotQueryServiceLeafStreamRetriever is a LeafStreamRetriever that retrieves a stream of [Leaf]s from the Hotshot Query Service. It expects the base URL of the Hotshot Query Service to be provided so that it can make the request to the Hotshot Query Service.

ProcessNodeIdentityUrlStreamTask

ProcessNodeIdentityUrlStreamTask is a task that processes a stream of Urls that are expected to contain a Node Identity. It will attempt to retrieve the Node Identity from the Url and then send it to the [Sink] provided.

ProcessProduceLeafStreamTask

ProcessProduceLeafStreamTask is a task that produce a stream of [Leaf]s from the Hotshot Query Service. It will attempt to retrieve the [Leaf]s from the Hotshot Query Service and then send them to the [Sink] provided.

PublishHotShotConfig

SequencerConfig

Enums

DefineApiError

EndpointError

Error

GetNodeIdentityFromUrlError

LoadApiError

LoadTomlError

RetrieveLeafStreamError 🔒

RetrieveLeafStreamError indicates the various failure conditions that can occur when attempting to retrieve a stream of [Leaf]s using the ProcessProduceLeafStreamTask::retrieve_leaf_stream function.

Constants

BASE_VERSION

Constant for the version of this API.

STATIC_VER_0_1

VERSION_0_1

VERSION_MAJ

CONSTANT for protocol major version

VERSION_MIN

CONSTANT for protocol minor version

Traits

LeafStreamRetriever

LeafStreamRetriever is a general trait that allows for the retrieval of a list of Leaves from a source. The specific implementation doesn’t care about the source, only that it is able to retrieve a stream of Leaves.

StateClientMessageSender

StateClientMessageSender allows for the retrieval of a [Sender] for sending messages received from the client to the Server for request processing.

Functions

define_api

get_node_identity_from_url

get_node_identity_from_url retrieves a NodeIdentity from a URL. It expects a url::Url to be provided so that it can make the request to the Sequencer status metrics API. It will return a NodeIdentity that is populated with the data retrieved from the Sequencer status metrics API. If no NodeIdentity is found, it will return a GetNodeIdentityFromUrlError::NoNodeIdentity error.

get_stake_table_from_sequencer

get_stake_table_from_sequencer retrieves the stake table from the Sequencer. It expects a [surf_disco::Client] to be provided so that it can make the request to the Hotshot Query Service. It will return a [StakeTable] that is populated with the data retrieved from the Hotshot Query Service.

load_api 🔒

node_identity_from_scrape

node_identity_from_scrape creates a NodeIdentity from a [Scrape]. It expects the [Scrape] to contain the necessary information to populate the NodeIdentity. If the [Scrape] doesn’t contain the necessary information to populate the NodeIdentity, then it will return None.

populate_node_identity_from_scrape

populate_node_identity_from_scrape populates a NodeIdentity from a [Scrape] that is expected to contain the necessary information to populate the NodeIdentity.

populate_node_identity_general_from_scrape 🔒

populate_node_identity_general_from_scrape populates the general information of a NodeIdentity from a [Sample] that is expected to be the “consensus_node_identity_general” sample.

populate_node_location_from_scrape 🔒

populate_node_location_from_scrape populates the location information of a NodeIdentity from a [Sample] that is expected to be the “consensus_node_identity_location” sample.

Type Aliases

Version01

Specific type for version 0.1