node_metrics::api::node_validator::v0

Module create_node_validator_api

Structs

NodeValidatorAPI

NodeValidatorConfig

SubmitPublicUrlsToScrapeTask

SubmitPublicUrlsToScrapeTask is a task that is capable of submitting public urls to a url sender at a regular interval. This task will submit the provided urls to the url sender every 5 minutes.

Enums

CreateNodeValidatorProcessingError

Constants

PUBLIC_URL_RESUBMIT_INTERVAL 🔒

Functions

create_node_validator_processing

create_node_validator_processing is a function that creates a node validator processing environment. This function will create a number of tasks that will be responsible for processing the data streams that are coming in from the various sources. This function will also create the data state that will be used to store the state of the network.