node_metrics::api::node_validator::v0::create_node_validator_api
Function create_node_validator_processing

pub async fn create_node_validator_processing(
    config: NodeValidatorConfig,
    internal_client_message_receiver: Receiver<InternalClientMessage<Sender<ServerMessage>>>,
    leaf_receiver: Receiver<Leaf<SeqTypes>>,
) -> Result<NodeValidatorAPI<Sender<Url>>, CreateNodeValidatorProcessingError>

create_node_validator_processing is a function that creates a node validator processing environment. This function will create a number of tasks that will be responsible for processing the data streams that are coming in from the various sources. This function will also create the data state that will be used to store the state of the network.