node_metrics::api::node_validator::v0

Function get_node_identity_from_url

pub async fn get_node_identity_from_url(
    url: Url,
) -> Result<NodeIdentity, GetNodeIdentityFromUrlError>

get_node_identity_from_url retrieves a NodeIdentity from a URL. It expects a url::Url to be provided so that it can make the request to the Sequencer status metrics API. It will return a NodeIdentity that is populated with the data retrieved from the Sequencer status metrics API. If no NodeIdentity is found, it will return a GetNodeIdentityFromUrlError::NoNodeIdentity error.