node_metrics::api::node_validator::v0

Trait UpdateBlockHeightForEntry

pub trait UpdateBlockHeightForEntry<T> {
    // Required methods
    fn block_height_for_entry(&self, entry: &T) -> u64;
    fn update_block_height_for_entry(&mut self, entry: &T);
}

UpdateBlockHeightForEntry is a trait that allows for the updating of the block height for a given entry. This is useful for updating the last received block height for a stream.

Required Methods

fn block_height_for_entry(&self, entry: &T) -> u64

[block_height_for_entry] returns the block height for the given entry. This is useful for updating the last received block height for a stream.

fn update_block_height_for_entry(&mut self, entry: &T)

[update_block_height_for_entry] updates the block height for the given entry. This is useful for updating the last received block height for a stream.