Function node_metrics::service::data_state::process_incoming_leaf

async fn process_incoming_leaf<BDSink, BVSink>(
    leaf: Leaf<SeqTypes>,
    data_state: Arc<RwLock<DataState>>,
    block_sender: BDSink,
    voters_sender: BVSink,
) -> Result<(), ProcessLeafError>
where
    Header: BlockHeader<SeqTypes> + QueryableHeader<SeqTypes> + ExplorerHeader<SeqTypes>,
    Payload: BlockPayload<SeqTypes>,
    BDSink: Sink<BlockDetail<SeqTypes>, Error = SendError> + Unpin,
    BVSink: Sink<BitVec<u16>, Error = SendError> + Unpin,

process_incoming_leaf is a helper function that will process an incoming [Leaf] and update the DataState with the new information. Additionally, the block that is contained within the [Leaf] will be computed into a [BlockDetail] and sent to the [Sink] so that it can be processed for real-time considerations.